Added a learning algorithm to RPS
Every time an input is repeated 3 (or more) times in a row, the computer adds 1 of its opposite move to the computer moves pool, thus making it more likely for the computer to win. However, the new "Reset Computer Choices on Player Input Change" button has been added to address the biggest flaw of this strategy: a player could spam rock (thus increasing the likelihood of paper) and then switch to scissors and win way too often. Now, if this button is checked, every time the player selects a new input (after 3 or more of another), the computer moves pool is reset to ["rock", "paper", "scissors"].